requirements.txt
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
Cython==3.0.5
fonttools==4.44.0
kiwisolver==1.4.5
matplotlib==3.8.1
numpy==1.26.2
opencv-python==4.8.1.78
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
pyparsing==3.1.1
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
tzdata==2023.3
